tabulated input improved to behave as a tabulator should
Tabulator input has suffered some modifications in order to match behaves with a terminal tabulator.
For instance, when setting a zero offset and the current cursor position lacks field, the operation is canceled.
In another case, if the current cursor position does not contain a field, but we've got an offset bigger than zero, it will lookup for the field to positions above.